Texture animation handled via ACS
The texture animation is now handled through the map ACS.  I had to take
this approach as previously - the animation would only occur on one side
of the sidedef but not both.  To resolve this issue, I created a small
script that will cycle through the animation as required on both sides
of the linedef.