Added fog and rain to the map
I have added fog and rain into the map.  The rain effect shouldn't be as
laggy on typical computers today, I tried to keep the rain spawners to a
minimal as the rain SPRITES are hardly visible in the map with sector
fog colour depth.  Moreover, the fog will help produce the dark and
dreary feel to the map.  This, also provides a nice and unintentional
light effect with the light tower, as the dynamic light hits the fog
sprite in such a way that the fog illuminates to a full bright.