Red teleport points to Death Trap
Due to time, I can't fully dedicate on making the secondary portion of
the map.  Sorry
If someone in the future wants to add on to this map or overhaul it,
they may do as they wish with it as long as proper credit is giving - no
permission is required by me.