Added Vid Compat. Check
Straight from TGRDM3; this script will try to detect the video renderer
and warn the user that the video renderer is incorrect if Software mode
was detected.  This map is OpenGL only and is not compatible with the
software renderer as of today and with today's Dev. version of ZDoom.