Added TGRDM3's Lightfall back mode
If incase the end-user is not using the OpenGL renderer or has point
lights disabled, this unique script will illuminate the entire map from
'128' to '192'.  This will provide users enough light to see what is
going within the map.  Of course, when Point Lights are enabled via
client settings - then the reverse occurs, the light will change from
'192' to '128' plus dynamic lighting.